driver: use dedicated SG lists for USERPTR buffers
The initial intent was to user the descriptor chain to pass USERPTR
buffers and buffers for large controls to the device, but this approach
hit two important limitations:

1. Large buffers can easily consume our entire set of descriptors,
   unless we set the size of the commandq to unreasonable sizes, and
   even there it is not guaranteed to scale nicely.
2. Buffers that are written by the device would need to be mapped in
   the device-writable area of the descriptor chain to be compliant with
   virtio's expectations, which complicates processing on the device
   side.

Alter the spec and driver code to instead use a classical SG-list mapped
to (typically) one descriptor instead. This is a bit less elegant, but
not really more complicated and allows us to pass buffers only once in
the device-readable section of the descriptor chain, since that
information is read-only for the device even though the buffer itself
might be writable.