use the guest physical address to unmap a mapped MMAP buffer
Update both the device and driver to use the guest physical address of a
mapped MMAP buffer in order to perform the MUNMAP command, instead of
its V4L2 offset.

Since buffer mappings can survive the buffers themselves, using the V4L2
offset means that offset must remain valid even after a buffer is
destroyed if there are active mappings. This clutters the offset address
space with buffers that cannot be mapped anymore, and complicates the
code.

The original munmap semantic also uses the user-space address to do the
unmapping, so on top of making things simpler this also brings us closer
to the original V4L2 behavior.