decoder/v4l2_device_proxy: use proxy epoll for session polling
The semantics of polling a V4L2 device were leaked into the public
SessionPoller trait, which required working with a V4l2Device and
enabling or disabling the polling of CAPTURE buffers. This was inelegant
and confusing, since this requirement is specific to the proxy device.

Improve this by moving the V4L2 specific parts into the proxy device,
and using a proxy epoll FD to signal the SessionPoller. This makes the
proxy device work with simpler semantics (the FD signals it is readable
when there are pending events) and allows us to use SessionPoller with
other devices as well.