device: add first version of the stateful decoder device
Add a device that implements the V4L2 stateful decoder specification. It
takes care of interpreting the V4L2 decoding protocol and delegates the
actual decoding work to a backend, which has a much simpler interface.

The device is still perfectible, but can do basic decoding flawlessly so
we will improve it in follow-up CLs.