device: move poller into runner structure
Now that we have harmonized polling between devices (a FD signals it is
readable if there is an event pending), we can move the session poll FDs
management out of each individual devices, and into the common runner
structure. This removes one generic parameter from the device structures
and will eventually result in less code.