device/v4l2_device_proxy: always try to export the buffer when doing MMAP
MMAP offsets are per-device, which means that if we are not careful we
might allow a session to map a buffer belonging to another session.
Avoid this by always exporting the FD of a requested buffer on the host.
Since we use the session to do that, the host kernel will deny access if
the buffer did not belong to the requesting session.