device/video_decoder: better crop rectangle handling
The crop rectangle was purely static until now, but this only works if
the encoded media contains format information and the client relies on
the initial DRC event to set up the CAPTURE buffers.

Strangely ffmpeg does not do that for H.264 and parses the SPS itself
to allocate CAPTURE buffers before even sending the first encoded chunk
to us.

This means that unless we let it set the CAPTURE coded size and crop
rectangle, it will work with the initial values for these, which will be
invalid.

Address this by letting the client set the crop rectangle itself as long
as we haven't received a DRC event.