device/mmap: allow overlapping buffers in the V4L2 MMAP address space
The range of V4L2 buffers inside the MMAP address space is actually
allowed to overlap. The only thing that matters is the start offset of
the buffer, which allows mmap to actually identify the buffer.

This makes sense considering that the MMAP address space is only 32
bits.

This restores functionality of proxied UVC camera devices on recent
kernels, which was broken due to the range of MMAP buffers overlapping.