protocol: make the MMAP command's offset parameter 32-bit
V4L2 uses a 32-bit unsigned integer for this, and since the buffers are
allowed to overlap in the MMAP address space as long as they have a
different start offset, we are not at risk of running out of space
anymore.

Aligning to the type used by V4L2 avoids some unneeded type conversions.