device: reorganize code for proper endianness handling
Virtio-media requires all data to transit in little-endian ordering, but
the current code absolutely ignored this, assuming both the host and
guest would be little-endian.

Fix this by introducing a VmediaType trait that all types transiting
through virtio must implement to provide little-endian conversion, and
changing the code to ensure that all access to virtio descriptors is
done with the little ordering conversion. This results in a rather big
change, but hopefully the code is also now simpler to grasp, as relevant
types have been grouped into the `io` module.

At the moment, build will fail if the host is big-endian - we need to
write proper big-endian conversion methods for types that can transit through
virtio-media and are still considering what the best way to do this
would be.