device/ffmpeg-decode: media crate in Cargo.toml
In order to add the latest published virtio-media
crate and "virtio-media-ffmpeg-decoder" as
dependencies simoultaneously, the latter cannot
rely on the upstream folder in its Cargo.toml.
Otherwise, the `VirtioDecoderBackend` will
differ and the package using them will fail
to build.

Instead, just use the published crate in
crates.io to be consistent.

Signed-off-by: Albert Esteve <aesteve@redhat.com>